Fix lag back with anticheats 1.17
