--- src/main/java/net/earthcomputer/multiconnect/protocols/v1_17/mixin/MixinClientPlayerInteractionManager.java
@@ -1,24 +1,0 @@
-package net.earthcomputer.multiconnect.protocols.v1_17.mixin;
-
-import net.earthcomputer.multiconnect.api.Protocols;
-import net.earthcomputer.multiconnect.impl.ConnectionInfo;
-import net.minecraft.client.network.ClientPlayNetworkHandler;
-import net.minecraft.client.network.ClientPlayerInteractionManager;
-import net.minecraft.network.Packet;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
-import org.spongepowered.asm.mixin.injection.Slice;
-
-@Mixin(value = ClientPlayerInteractionManager.class, priority = 0)
-public class MixinClientPlayerInteractionManager {
-
-    @Redirect(method = "interactItem", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/network/ClientPlayNetworkHandler;sendPacket(Lnet/minecraft/network/Packet;)V", ordinal = 0),
-            slice = @Slice(from = @At(value = "INVOKE", target = "Lnet/minecraft/client/network/ClientPlayerInteractionManager;syncSelectedSlot()V"), to = @At(value = "INVOKE", target = "Lnet/minecraft/client/network/ClientPlayNetworkHandler;sendPacket(Lnet/minecraft/network/Packet;)V", ordinal = 1)))
-    private void redirectPlayerPosPacket(ClientPlayNetworkHandler clientPlayNetworkHandler, Packet<?> packet) {
-        if(ConnectionInfo.protocolVersion >= Protocols.V1_17){
-            clientPlayNetworkHandler.sendPacket(packet);
-        }
-    }
-
-}
